# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RSSwiper",
    "version": "1.0",
    "description": "RSSwiper is a simple RSS reader that allows you to read your favorite newsfeeds in a swipeable interface. It is available as a web app. \n\nTo ensure that all team members are on the same page and that the development follows a common thread, this documentation of the RSSwiper server has been created according to the OpenAPI standard.\n\nFor more information about RSSwiper, visit the [GitLab repository](https://git.mylab.th-luebeck.de/vwprg/wise23_24/b2)."
  },
...
  "servers": [
    {
      "url": "{protocol}://{host}:{port}/{endpoint}",
      "description": "Development server",
      "variables": {
        "protocol": {
          "enum": [
            "http",
            "https"
          ],
          "default": "http"
        },
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "8080"
        },
        "endpoint": {
          "default": "v1"
        }
      }
    },
    {
      "url": "https://backend.rsswipe.mcloud.digital/v1",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/articles": {
      "get": {
        "description": "Get the newest unseen articles of your feeds",
        "tags": [
          "article"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "The maximum number of articles to return",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1
            }
          },
          {
            "name": "feeds",
            "in": "query",
            "description": "The ids of the feeds to return articles from. If not set, articles from all feeds are returned",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "categories",
            "in": "query",
            "description": "The ids of the categories to return articles from. If not set, articles from all categories are returned",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "The date from which (startDate included) articles should be returned. If not set, articles from all dates are returned",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "The date until which (endDate excluded) articles should be returned. If not set, articles from all dates are returned",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The newest unseen articles of all or selected feeds",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidInputError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/articles/{id}": {
      "get": {
        "description": "Get an article by id",
        "tags": [
          "article"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the article",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The article",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidInputError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "put": {
        "description": "Mark an article as seen or read",
        "tags": [
          "article"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the article",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Article marked successfully"
          },
          "400": {
            "$ref": "#/components/responses/InvalidInputError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/articles/{id}/content": {
      "get": {
        "description": "Get the content of an article",
        "tags": [
          "article"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the article",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The content of the article",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleContent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidInputError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
    "/readinglist/articles": {
      "post": {
        "description": "Add an article to your readinglist",
        "tags": [
          "readinglist"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "The id of the article"
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article added to readinglist"
          },
          "400": {
            "$ref": "#/components/responses/InvalidInputError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "delete": {
        "description": "Remove an article from your readinglist",
        "tags": [
          "readinglist"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "The id of the article"
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article removed from readinglist"
          },
          "400": {
            "$ref": "#/components/responses/InvalidInputError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
    "/starred/articles": {
      "post": {
        "description": "Star an article",
        "tags": [
          "starred"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "The id of the article"
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article starred"
          },
          "400": {
            "$ref": "#/components/responses/InvalidInputError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "delete": {
        "description": "Unstar an article",
        "tags": [
          "starred"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "The id of the article"
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article unstarred"
          },
          "400": {
            "$ref": "#/components/responses/InvalidInputError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
    "/articles/content": {
      "post": {
        "description": "Get the stored content of multiple articles. Ids without successfully extracted content map to null. No live extraction happens here.",
        "tags": [
          "article"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticlesContentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Map of article id to its content (same shape as the single content endpoint) or null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticlesContentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidInputError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "title": {
            "type": "string"
          },
          "link": {
            "type": "string",
            "format": "uri"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "read": {
            "type": "boolean"
          },
          "seen": {
            "type": "boolean"
          },
          "feed": {
            "$ref": "#/components/schemas/Feed"
          }
        }
      },
      "UpdateArticleInput": {
        "type": "object",
        "properties": {
          "read": {
            "type": "boolean"
          },
          "seen": {
            "type": "boolean"
          },
          "starred": {
            "type": "boolean"
          },
          "saved": {
            "type": "boolean"
          }
        },
        "required": [
          "id"
        ]
      },
      "ArticleContent": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "article title"
          },
          "content": {
            "type": "string",
            "description": "HTML string of processed article content"
          },
          "textContent": {
            "type": "string",
            "description": "text content of the article, with all the HTML tags removed"
          },
          "length": {
            "type": "integer",
            "description": "length of an article, in characters"
          },
          "excerpt": {
            "type": "string",
            "description": "article description, or short excerpt from the content"
          },
          "byline": {
            "type": "string",
            "description": "author metadata"
          },
          "dir": {
            "type": "string",
            "description": "content direction"
          },
          "siteName": {
            "type": "string",
            "description": "name of the site"
          },
          "lang": {
            "type": "string",
            "description": "content language"
          }
        }
      },
...
      "Feed": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "link": {
            "type": "string",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "faviconUrl": {
            "type": "string",
            "format": "uri"
          },
          "lastUpdate": {
            "type": "string",
            "format": "date-time"
          },
          "openInApp": {
            "type": "boolean"
          },
          "errormessage": {
            "type": "string",
            "description": "the error message if there was an error parsing the feed"
          },
          "description": {
            "type": "string",
            "description": "the description of the feed"
          },
          "error_count": {
            "type": "number",
            "description": "the number of times the parsing of the feed failed since the last success"
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ArticlesContentInput": {
        "type": "object",
        "required": [
          "ids"
        ],
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "description": "Article ids, at most CONTENT_PREFETCH_MAX_IDS"
          }
        }
      },
      "ArticlesContentResponse": {
        "type": "object",
        "additionalProperties": {
          "oneOf": [
            {
              "$ref": "#/components/schemas/ArticleContent"
            },
            {
              "type": "null"
            }
          ]
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      },
      "InvalidInputError": {
        "description": "Invalid input",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}